feat: add number formatting utilities
